requirements.txt
aiohttp==3.6.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
braceexpand==0.1.6
chardet==3.0.4
discord.py==1.6.0
humanize==3.0.1
idna==2.10
import-expression==1.1.4
jishaku==1.20.0.220
multidict==4.7.6
six==1.15.0
yarl==1.5.1